requirements.txt
mediapipe==0.8.11
opencv-python==4.5.3.56
pyserial==3.5